+++ backends/rapidpro/msg.go
@@ -18,7 +18,7 @@ import (
 
 	"mime"
 
-	"github.com/garyburd/redigo/redis"
+	"github.com/gomodule/redigo/redis"
 	"github.com/lib/pq"
 	"github.com/nyaruka/courier"
 	"github.com/nyaruka/courier/queue"
@@ -193,12 +193,36 @@ WHERE
 	id = $1
 `
 
+const selectChannelSQL = `
+SELECT
+	org_id,
+	ch.id as id,
+	ch.uuid as uuid,
+	ch.name as name,
+	channel_type, schemes,
+	address, role,
+	ch.country as country,
+	ch.config as config,
+	org.config as org_config,
+	org.is_anon as org_is_anon
+FROM
+	channels_channel ch
+	JOIN orgs_org org on ch.org_id = org.id
+WHERE
+    ch.id = $1
+`
+
 // for testing only, returned DBMsg object is not fully populated
 func readMsgFromDB(b *backend, id courier.MsgID) (*DBMsg, error) {
 	m := &DBMsg{
 		ID_: id,
 	}
 	err := b.db.Get(m, selectMsgSQL, id)
+	ch := &DBChannel{
+		ID_: m.ChannelID_,
+	}
+	err = b.db.Get(ch, selectChannelSQL, m.ChannelID_)
+	m.channel = ch
 	return m, err
 }