+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+v6.0.0
+----------
+ * Update README
+
+v5.7.12
+----------
+ * URN channel change only for channels with SEND role
+ * Update to gocommon v1.6.1
+ * Add RocketChat handler
+ * Add discord handler
+
+v5.7.11
+----------
+ * Cache media ids for WhatsApp attachments
+
 v5.7.10
 ----------
  * Support receiving Multipart form data requests for EX channels